organized server files
