+++ app/src/components/auth/signin/signin.html
@@ -3,13 +3,13 @@
 	<form name="auth" ng-submit="$ctrl.authenticate()">
 		<div>
 			<label>
-				username: <input ng-model="$ctrl.credentials.username">
+				Username: <input ng-model="$ctrl.credentials.username">
 			</label>
 		</div>
 
 		<div>
 			<label>
-				password: <input type="password" ng-model="$ctrl.credentials.password">
+				Password: <input type="password" ng-model="$ctrl.credentials.password">
 			</label>
 		</div>
