+++ server/lib/cors.js
@@ -1,0 +1,9 @@
+module.exports = function cors(url){
+  return (req, res, next)=>{
+    const url = '*';
+    res.header('Access-Control-Allow-Origin', url);
+    res.header('Access-Control-Allow-Methods', 'POST, GET, OPTIONS, PUT, DELETE');
+    res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Authorization');
+    next();
+  };
+};
