+++ server/lib/ensureAuth.js
