+++ server/lib/token.js
