+++ server/models/user.js
@@ -4,6 +4,7 @@ const bcrypt = require('bcrypt');
 
 const userSchema = new Schema({
   username: {type: String, required: true},
+  //TODO: add in email match regex here
   email: {type: String},
   password: {type: String, required: true},
   roles: [String]
